requirements.txt
black==24.10.0
bpython
celery==5.4.0
Django==4.2
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
flake8==7.1.1
gunicorn==23.0.0
isort==5.13.2
pre-commit==4.0.1
psycopg2==2.9.10
python-decouple==3.8
Pillow
